+++ main1263.cpp
@@ -30,7 +30,7 @@ int main()
  e9_Cpm_4n4 = (unsigned short)687847132;
  Pi7W6 = (unsigned short)-2024901123;
  PZtB4Z = (unsigned short)1216246903;
- XWgt = -1722603867;
+ XWgt = (unsigned short)-1722603867;
  RjcxW = (unsigned short)-701616175;
  EeA_37SSn7 = (unsigned short)1615153093;
  gR_wslI = (unsigned short)86673499;
@@ -40,7 +40,7 @@ int main()
  umM74lP = (unsigned short)796815960;
  O2ZaZdXp = (unsigned short)-1685688513;
  TQ_U__A6d = (unsigned short)-1357701171;
- owfm7u008f = -1244326728;
+ owfm7u008f = (unsigned short)-1244326728;
  W7_86_SV = (unsigned short)231229966;
  cC6dRTVUH1 = (unsigned short)53361747;
  iTvV = (unsigned short)1663260594;
@@ -50,7 +50,7 @@ int main()
  y8b34oe = (unsigned short)1490844708;
  S66U3vR = (unsigned short)-626099581;
  K_W828Aw9 = (unsigned short)1996279813;
- qd9Si_ = 668164761;
+ qd9Si_ = (unsigned short)668164761;
  E02tHQ36 = (unsigned short)-1927654509;
  EWZQZ_YOa = (unsigned short)-1029480401;
  lVa1 = (unsigned short)575766363;
@@ -60,7 +60,7 @@ int main()
  ghr5 = (unsigned short)-349885556;
  Jr8PZw = (unsigned short)-1728776128;
  eh_g = (unsigned short)-802191269;
- F_3aeQg7 = -936440368;
+ F_3aeQg7 = (unsigned short)-936440368;
  lTh_dO6 = (unsigned short)-1434297192;
  aoa27Ao9f4 = (unsigned short)1830573474;
  N2xHK3ydN = (unsigned short)369177018;
@@ -70,7 +70,7 @@ int main()
  y9r4A = (unsigned short)-1539561205;
  A8PGdjDWg0 = (unsigned short)255212206;
  T747H_tXZ0 = (unsigned short)-1804266307;
- rC1By0Q = 415279185;
+ rC1By0Q = (unsigned short)415279185;
  d92UeB_fdD = (unsigned short)664566199;
  w09V_l7 = (unsigned short)1116489376;
  HwEwh7RL = (unsigned short)-846776431;
